implement clean architecture, follow strict dart:lint rules
