--- lib/custom_widgets.dart
@@ -1,6 +1,0 @@
-export './../../widgets/genre_list.dart';
-export './../../widgets/like_counter.dart';
-export './../../widgets/movie_overview_text.dart';
-export './../../widgets/details_poster_section.dart';
-export './../../widgets/details_title_section.dart';
-export './../../widgets/details_backdrop_section.dart';
