+++ lib/main.dart
@@ -1,16 +1,34 @@
 import 'package:flutter/material.dart';
-import '../../../pages/movies_page.dart';
-import 'utils/movie_repository.dart';
 
-void main() => runApp(const MyApp());
+import 'src/config/theme/app_themes.dart';
+import 'src/data/repository/movie_repository.dart';
+import 'src/domain/repository/i_movie_repository.dart';
+import 'src/presentation/view/movies_view.dart';
+
+void main() {
+  final IMovieRepository repository = MovieRepository();
+  runApp(
+    MyApp(
+      repository: repository,
+    ),
+  );
+}
 
 class MyApp extends StatelessWidget {
-  const MyApp({super.key});
+  const MyApp({
+    super.key,
+    required this.repository,
+  });
+
+  final IMovieRepository repository;
 
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      home: MoviesPage(repository: MovieRepository()),
+      theme: AppTheme.dark,
+      home: MoviesView(
+        repository: repository,
+      ),
     );
   }
 }
