+++ lib/src/core/util/ui_constants.dart
@@ -1,0 +1,18 @@
+import 'dart:ui' show Color;
+
+abstract class UIConstants {
+
+  ///Details page Title section
+  static const Color titleSectionTextColor = Color(0xFFFFFFFF);
+
+  ///Details page Poster section
+  static const double posterSectionIconSize = 120;
+
+  ///Details page Overview section
+  static const Color overviewSectionTextColor = Color(0x99FFFFFF);
+
+  ///Genre Card styling
+  static const double genreCardBorderRadius = 3;
+  static const double genreCardBorderWidth = 2;
+  static const Color genreCardTextColor = Color(0xB3FFFFFF);
+}
