+++ lib/src/domain/repository/i_movie_repository.dart
@@ -1,0 +1,5 @@
+import '../entity/movie.dart';
+
+abstract interface class IMovieRepository {
+  Future<List<Movie>?> getMovies();
+}
