--- lib/utils/i_movie_repository.dart
@@ -1,5 +1,0 @@
-import '../../../models/movie.dart';
-
-abstract interface class IMovieRepository {
-  Future<List<Movie>?> getMovies();
-}
