+++ pubspec.yaml
@@ -10,29 +10,26 @@ environment:
   sdk: '>=3.0.6 <4.0.0'
 
 dependencies:
+  cupertino_icons: ^1.0.2
   flutter:
     sdk: flutter
-  sqflite:
   provider: ^6.0.1
-
-
-  cupertino_icons: ^1.0.2
   shared_preferences: ^2.2.0
+  sqflite:
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  flutter_lints: ^2.0.0
+  lint: ^2.0.0
 
 flutter:
 
   uses-material-design: true
 
   assets:
-    - images/
     - mock_data/
-
+    - images/
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware
