+++ test/widget_test.dart
@@ -9,11 +9,17 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'package:movie_app_jsalinas/main.dart';
+import 'package:movie_app_jsalinas/src/data/repository/movie_repository.dart';
 
 void main() {
   testWidgets('Counter increments smoke test', (WidgetTester tester) async {
     // Build our app and trigger a frame.
-    await tester.pumpWidget(const MyApp());
+    final repository = MovieRepository();
+    await tester.pumpWidget(
+      MyApp(
+        repository: repository,
+      ),
+    );
 
     // Verify that our counter starts at 0.
     expect(find.text('0'), findsOneWidget);
